+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # 版本更新 
 
 ## v0.2
+### 2021/06/11
+- update code style of the imgMode component and added `virtualHost` field
+- add scaleToFill mode to imgMode component
+
 ### 2021/06/09
 - add heightFix mode to imgMode component